--- hw4_advance_algorithm/export/trie.json
@@ -1,1 +1,0 @@
-{"index":".","isEndWord":false,"content":"root"}
