--- hw4_advance_algorithm/resource/output.json
@@ -1,22 +1,0 @@
-[
-  {
-    "pred_no_corret": "X. Thuận Thành, H. Cần Giuộc, T. Long An",
-    "district": "Binh Thuan"
-  },
-  {
-    "pred_no_corret": "Thuận Thanh, HCần Giuộc, Tlong An",
-    "district": "Binh Thuan"
-  },
-  {
-    "pred_no_corret": "Thuận Thành, H Cần Giuộc T. Long An",
-    "district": "Binh Thuan"
-  },
-  {
-    "pred_no_corret": "X Thuận Thành H. Cần Giuộc, Long An",
-    "district": "Binh Thuan"
-  },
-  {
-    "pred_no_corret": "X ThuanThanh H. Can Giuoc, Long An",
-    "district": "Binh Thuan"
-  }
-]
