+++ hw4_advance_algorithm/src/JsonAddressObject.java
@@ -1,40 +1,27 @@
 package src;
 
 public class JsonAddressObject {
-    private String pred_no_corret;
-    private String province;
-    private String district;
-    private String ward;
+    private String text;
 
-    public String getPred_no_corret() {
-        return pred_no_corret;
+    private JsonAddressNestedObject address;
+
+    public JsonAddressObject(JsonAddressNestedObject address) {
+        this.address = address;
     }
 
-    public void setPred_no_corret(String pred_no_corret) {
-        this.pred_no_corret = pred_no_corret;
+    public String getText() {
+        return text;
     }
 
-    public String getProvince() {
-        return province;
+    public void setText(String text) {
+        this.text = text;
     }
 
-    public void setProvince(String province) {
-        this.province = province;
+    public JsonAddressNestedObject getAddress() {
+        return address;
     }
 
-    public String getDistrict() {
-        return district;
-    }
-
-    public void setDistrict(String district) {
-        this.district = district;
-    }
-
-    public String getWard() {
-        return ward;
-    }
-
-    public void setWard(String ward) {
-        this.ward = ward;
+    public void setAddress(JsonAddressNestedObject address) {
+        this.address = address;
     }
 }
